Revert "Give filter masks a non-overlay sprite."
This reverts commit a9825f9e490ad7f356fc0e4ad6d9fe838cc07b09. Acepleiades
rightly pointed out in the comments of #928 that this change could be
confusing for whoever eventually implements a unique filter mask sprite.

For the sake of practice, I'm filing this PR and adding a `looks_like`
entry to achieve the same results in the main repository.

Sorry for the extra paperwork!