Added Black Dragon files (#2532)
* Added Black Dragon files

Added sprite sheet, json, and credits files for black dragon mutation

* Update magiclysm_black_dragon.json

Corrected error where the dragon legs mutation had male set twice and female was missing

* Re-configuration of sprites

Re-configuration of sprites to match the rest of the repo. They're handled differently in CDDA's base tileset download, so I was confused as to how the layout was supposed to work at first.