[NeoDays] fix some wielded gun ids
With the addition of modular AR-15 and M4 rifles, the IDs got changed
but not updatd here, resulting in missing sprites when wielding those
guns.

I'm not too familiar with this repo, but AFAICT this fix the issue - I
was able to generate the tileset and load on a local build of CDDA. This
fix may not be exhaustive - i.e. there may be further guns with old IDs.